New feature: Stream new messages in TCP mode
When giving the `-T` flag, historical behavior is that the previously
received, buffered messages are streamed to the client, and then no
more. The socket is kept open until either the client closes it, or until
the client sends any data, at which point the client is disconnected.

When the `-k` flag is given in addition to the `-T` flag, an alternative
behavior is now implemented: As new messages are decoded, they
are sent to any attached clients in realtime. The client does not
need to disconnect to continue receiving updates.

Closes #33